+++ CMakeLists.txt
@@ -1,10 +1,37 @@
-cmake_minimum_required(VERSION 3.16.3)
+cmake_minimum_required(VERSION 3.5)
 
 project(binlex)
 
-add_executable(${PROJECT_NAME} main.cpp)
+set(SOURCES_BINLEX
+    src/binlex.cpp
+    src/args.cpp
+    src/raw.cpp
+    src/common.cpp
+    src/blelf.cpp
+    src/pe.cpp
+    src/decompiler.cpp
+)
 
-# This is a hack (remove this later)
-#set(CMAKE_CXX_FLAGS "-fpermissive")
+add_executable(binlex ${SOURCES_BINLEX})
 
-target_link_libraries(${PROJECT_NAME} -lcapstone -lcrypto -lm)
+target_link_libraries(binlex -lcapstone -lcrypto -lm)
+
+target_include_directories(binlex
+    PRIVATE
+        ${PROJECT_SOURCE_DIR}/include
+)
+
+project(blyara)
+
+set(SOURCES_BLYARA
+    src/blyara.cpp
+)
+
+add_executable(blyara ${SOURCES_BLYARA})
+
+target_link_libraries(blyara -lcapstone -lcrypto -lm)
+
+target_include_directories(blyara
+    PRIVATE
+        ${PROJECT_SOURCE_DIR}/include
+)
