+++ src/binlex.cpp
@@ -1,43 +1,34 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <elf.h>
-#include <iostream>
-#include <string>
-#include <algorithm>
-#include "src/elf.h"
-#include "src/pe.h"
-#include "src/raw.h"
-#include "src/macho.h"
-#include "src/decompiler.h"
-#include "src/decompiler_rev.h"
-#include "src/args.h"
-#include "src/cil.h"
-#include "src/json.h"
-
-using json = nlohmann::json;
-using namespace std;
+#include <string.h>
+#include <capstone/capstone.h>
+#include "args.h"
+#include "decompiler.h"
+#include "raw.h"
+#include "pe.h"
+#include "blelf.h"
+#include "common.h"
 
 int main(int argc, char **argv){
     Args args;
+    Common common;
     args.parse(argc, argv);
+    if (args.options.mode == NULL){
+        args.print_help();
+        return 1;
+    }
     if (strcmp(args.options.mode, (char *)"elf:x86_64") == 0 &&
         args.options.io_type == ARGS_IO_TYPE_FILE){
-        Elf elf64;
-
-        if (elf64.Setup(ELF_MODE_X86_64) == false){
+        Elf elfx86_64;
+        if (elfx86_64.Setup(ELF_MODE_X86_64) == false){
             return 1;
         }
-        if (elf64.ReadFile(args.options.input) == false){
-            return 1;
-        }
-        if (elf64.GetExecutableData() == false){
+        if (elfx86_64.ReadFile(args.options.input) == false){
             return 1;
         }
         Decompiler decompiler;
         decompiler.Setup(CS_ARCH_X86, CS_MODE_64);
         for (int i = 0; i < ELF_MAX_SECTIONS; i++){
-            if (elf64.sections[i].data != NULL){
-                decompiler.x86_64(elf64.sections[i].data, elf64.sections[i].size, elf64.sections[i].offset, i);
+            if (elfx86_64.sections[i].data != NULL){
+                decompiler.x86_64(elfx86_64.sections[i].data, elfx86_64.sections[i].size, elfx86_64.sections[i].offset, i);
             }
         }
         if (args.options.output == NULL){
@@ -49,22 +40,18 @@ int main(int argc, char **argv){
     }
     if (strcmp(args.options.mode, (char *)"elf:x86") == 0 &&
         args.options.io_type == ARGS_IO_TYPE_FILE){
-        int result = false;
-        Elf elf32;
-        if (elf32.Setup(ELF_MODE_X86) == false){
+        Elf elfx86;
+        if (elfx86.Setup(ELF_MODE_X86) == false){
             return 1;
         }
-        if (elf32.ReadFile(args.options.input) == false){
-            return 1;
-        }
-        if (elf32.GetExecutableData() == false){
+        if (elfx86.ReadFile(args.options.input) == false){
             return 1;
         }
         Decompiler decompiler;
         decompiler.Setup(CS_ARCH_X86, CS_MODE_32);
         for (int i = 0; i < ELF_MAX_SECTIONS; i++){
-            if (elf32.sections[i].data != NULL){
-                decompiler.x86_64(elf32.sections[i].data, elf32.sections[i].size, elf32.sections[i].offset, i);
+            if (elfx86.sections[i].data != NULL){
+                decompiler.x86_64(elfx86.sections[i].data, elfx86.sections[i].size, elfx86.sections[i].offset, i);
             }
         }
         if (args.options.output == NULL){
@@ -76,7 +63,6 @@ int main(int argc, char **argv){
     }
     if (strcmp(args.options.mode, (char *)"pe:x86") == 0 &&
         args.options.io_type == ARGS_IO_TYPE_FILE){
-        int result = false;
         Pe pe32;
         if (pe32.Setup(PE_MODE_X86) == false){
             return 1;
@@ -100,7 +86,6 @@ int main(int argc, char **argv){
     }
     if (strcmp(args.options.mode, (char *)"pe:x86_64") == 0 &&
         args.options.io_type == ARGS_IO_TYPE_FILE){
-        int result = false;
         Pe pe64;
         if (pe64.Setup(PE_MODE_X86_64) == false){
             return 1;
@@ -126,14 +111,14 @@ int main(int argc, char **argv){
         args.options.io_type == ARGS_IO_TYPE_FILE){
         Raw rawx86;
         rawx86.ReadFile(args.options.input, 0);
-        DecompilerREV decompiler;
+        Decompiler decompiler;
         decompiler.Setup(CS_ARCH_X86, CS_MODE_32);
-        decompiler.Decompile(rawx86.sections[0].data, rawx86.sections[0].size, rawx86.sections[0].offset, 0);
-        // if (args.options.output == NULL){
-        //     decompiler.PrintTraits(args.options.pretty);
-        // } else {
-        //     decompiler.WriteTraits(args.options.output, args.options.pretty);
-        // }
+        decompiler.x86_64(rawx86.sections[0].data, rawx86.sections[0].size, rawx86.sections[0].offset, 0);
+        if (args.options.output == NULL){
+            decompiler.PrintTraits(args.options.pretty);
+        } else {
+            decompiler.WriteTraits(args.options.output, args.options.pretty);
+        }
         return 0;
     }
     if (strcmp(args.options.mode, (char *)"raw:x86_64") == 0 &&
@@ -152,28 +137,6 @@ int main(int argc, char **argv){
     }
     if (strcmp(args.options.mode, (char *)"raw:cil") == 0 &&
         args.options.io_type == ARGS_IO_TYPE_FILE){
-        Raw raw_cil;
-        if (raw_cil.ReadFile(args.options.input, 0) == false){
-            return 1;
-        }
-        CILDecompiler cil_decompiler;
-        if (cil_decompiler.Setup(CIL_DECOMPILER_TYPE_BLCKS) == false){
-            return 1;
-        }
-        if (cil_decompiler.Decompile(raw_cil.sections[0].data, raw_cil.sections[0].size, 0) == false){
-            return 1;
-        }
-        if (cil_decompiler.Setup(CIL_DECOMPILER_TYPE_FUNCS) == false){
-            return 1;
-        }
-        if (cil_decompiler.Decompile(raw_cil.sections[0].data, raw_cil.sections[0].size, 0) == false){
-            return 1;
-        }
-        if (args.options.output == NULL){
-            cil_decompiler.PrintTraits();
-        } else {
-            cil_decompiler.WriteTraits(args.options.output);
-        }
         return 0;
     }
     args.print_help();
